requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
einops==0.8.0
hydra-core==1.3.2
llava==1.2.2.post1
ms-swift==2.5.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.35.9
opencv-python==4.10.0.84
Pillow==10.3.0
PyYAML==6.0.1
requests==2.32.3
sentencepiece==0.2.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
